requirements.txt
zeep==1.6.0
flask==0.12
xmltodict==0.10.2
connexion==1.1.9
signxml==2.5.2
flex
click==6.7
werkzeug
lxmlbind
cryptography==2.2.2